refactoring to modules folder system, finished chat module, starting displayImages
