+++ vite-project/src/Modules/Chat/WebSocketChat/pages/WebSocketChat.tsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { useAuth0 } from "@auth0/auth0-react";
 import { Users, useWebSocketChat, User } from "../hooks/useWebSocketChat";
-import { useModal } from "../hooks/useModal";
-import Modal from "../components/Modal";
-import ChatModalContent from "../components/ModalComponents/ChatModalContent";
-import { SideBar } from "../components/ChatWebSockets/SideBar";
-import { MessageArea } from "../components/ChatWebSockets/MessageArea";
-import { ChatInputArea } from "../components/ChatWebSockets/ChatInputArea";
-import { useSharePost } from "../hooks/useSharePost";
-import Button from "../components/Button";
+import { useModal } from "../../../../hooks/useModal";
+import Modal from "../../../../components/Modal";
+import ChatModalContent from "../../../../components/ModalComponents/ChatModalContent";
+import { SideBar } from "../../Components/SideBar";
+import { MessageArea } from "../../Components/MessageArea";
+import { ChatInputArea } from "../../Components/ChatInputArea";
+import { useSharePost } from "../../../../hooks/useSharePost";
+import Button from "../../../../components/Button";
 
 const sortUsers = (userList: Users, chatUser: string): User[] => {
   const result = Object.keys(userList)
