+++ vite-project/src/Modules/DisplayImages/lib/types.ts
