+++ vite-project/src/components/index.ts
@@ -1,5 +1,4 @@
-import SinglePhotoCard from "./SinglePhotoCard";
 import FormField from "./FormField";
 import Loader from "./Loader";
 
-export { SinglePhotoCard, FormField, Loader };
+export { FormField, Loader };
