+++ vite-project/src/pages/index.ts
@@ -1,4 +1,4 @@
-import Home from "./Home";
+import Home from "../Modules/DisplayImages/pages/Home";
 import CreatePost from "./CreatePost";
 
 export { Home, CreatePost };
